fix git build some bug
